+++ netcore/src/OPADotNet.Core/Ast/Explanation/ExplanationExpression.cs
@@ -14,25 +14,6 @@ namespace OPADotNet.Core.Ast.Explanation
         [JsonPropertyName("node")]
         public new AstExpression Node { get; set; }
 
-        public override bool Equals(object obj)
-        {
-            return obj is ExplanationExpression body &&
-                   Operation == body.Operation &&
-                   QueryId == body.QueryId &&
-                   ParentId == body.ParentId &&
-                   Type == body.Type &&
-                   Equals(Node, body.Node) &&
-                   Message == body.Message &&
-                   Locals.AreEqual(body.Locals) &&
-                   Equals(Location, body.Location) &&
-                   Equals(Node, body.Node);
-        }
-
-        public override int GetHashCode()
-        {
-            return base.GetHashCode();
-        }
-
         private protected override AstNode GetNode() => Node;
     }
 }
