+++ netcore/src/OPADotNet.Core/Ast/Explanation/ExplanationNode.cs
@@ -1,4 +1,5 @@
 ﻿using OPADotNet.Ast.Models;
+using OPADotNet.Core.Extensions;
 using System;
 using System.Collections.Generic;
 using System.Text;
@@ -32,6 +33,20 @@ namespace OPADotNet.Core.Ast.Explanation
 
         [JsonPropertyName("location")]
         public ExplanationLocation Location { get; set; }
+
+        public override bool Equals(object obj)
+        {
+            return obj is ExplanationNode body &&
+                   Operation == body.Operation &&
+                   QueryId == body.QueryId &&
+                   ParentId == body.ParentId &&
+                   Type == body.Type &&
+                   Equals(Node, body.Node) &&
+                   Message == body.Message &&
+                   Locals.AreEqual(body.Locals) &&
+                   Equals(Location, body.Location) &&
+                   Equals(Node, body.Node);
+        }
 
         public override int GetHashCode()
         {
